requirements.txt
pandas==2.2.2
numpy==1.26.4
prophet==1.1.5
openai==1.30.1
streamlit==1.35.0
fpdf2==2.7.9
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
plotly==5.22.0
scikit-learn==1.5.0
pytest==8.2.2
